Fix missing Microsoft.Bcl.Hash reference
The CoseSign1 libraries were originally built for .NET Standard 2.1 and later downgraded to 2.0 for compatibility. Microsoft.Bcl.Hash is included in 2.1 but not 2.0, so with the downgrade, certain validation scenarios started to fail because they couldn't find Microsoft.Bcl.Hash.
This PR adds the assembly as a dependence of CoseSign1.Abstractions, which fixes the broken scenario.
It also fixes some usage text that was appearing incorrectly, or in some cases not being displayed at all.